Change the order of execution strategy and remove app.py file since s2i-python-container's readme provided wrong information.
