+++ .s2i/environment
@@ -1,20 +1,22 @@
 # Environment variables
 
 UPGRADE_PIP_TO_LATEST=TRUE
-APP_CONFIG=configs/production.py
 
 # Execution Strategy: Execute by following precedence order.
 
-# 1. Gunicorn
-# Search "wsgi.py" from repo or specified by
+# 1. Shell Script
 
-# APP_MODULE=MODULE_NAME:VARIABLE_NAME
+# APP_SCRIPT=app.sh
 
 # 2. Python Script
 
 # APP_FILE=app.py
 
-# 3. Shell Script
+# 3. Gunicorn
+# Search "wsgi.py" from repo or specified by
 
-# APP_SCRIPT=app.sh
+# APP_MODULE=MODULE_NAME:VARIABLE_NAME
+APP_CONFIG=configs/production.py
+
+
