+++ wsgi.py
@@ -1,3 +1,12 @@
-import app
+from flask import Flask
+from apis import api
 
-application = app.create_app()
+
+def create_app():
+    app = Flask(__name__)
+    app.config['RESTPLUS_MASK_SWAGGER'] = False
+    api.init_app(app)
+    return app
+
+
+application = create_app()
